`GridItem.tsx`, `ListItem.tsx` - rename
