+++ packages/@uppy/provider-views/src/Item/components/ListItem.tsx
